android 5.0 support

+++ src/com/mykingdom/downloader/DownloadFile.java
@@ -1,18 +1,24 @@
 package com.mykingdom.downloader;
 
-import java.io.BufferedInputStream;
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.InputStream;
-import java.io.OutputStream;
+import java.net.HttpURLConnection;
 import java.net.URL;
-import java.net.URLConnection;
-import java.util.ArrayList;
+
+import javax.net.ssl.SSLContext;
+import javax.net.ssl.SSLSocketFactory;
+import javax.net.ssl.TrustManager;
+import javax.net.ssl.X509TrustManager;
+
+import org.appcelerator.kroll.common.Log;
+
+import java.security.SecureRandom;
+import java.security.cert.X509Certificate;
 import java.util.HashMap;
 import java.util.List;
-
-import org.appcelerator.kroll.KrollDict;
-import org.appcelerator.titanium.TiFileProxy;
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
 
 import android.content.Context;
 import android.os.AsyncTask;
@@ -20,7 +26,7 @@ import android.os.AsyncTask;
 class DownloadFile extends AsyncTask<List<Object>, Integer, String> {
 
 	private File outputDirectory;
-	private boolean enableNotification = false;
+	private boolean enableNotification, useCache;
 	private Integer notificationId;
 	private String notificationTitle;
 	private Context mContext;
@@ -28,12 +34,13 @@ class DownloadFile extends AsyncTask<List<Object>, Integer, String> {
 	IAsyncFetchListener fetchListener = null;
 
 	public DownloadFile(Context context, File direcory, boolean enable,
-			Integer id, String title) {
+			Integer id, String title, boolean cache) {
 		mContext = context;
 		outputDirectory = direcory;
 		enableNotification = enable;
 		notificationId = id;
 		notificationTitle = title;
+		useCache = cache;
 	}
 
 	public void setListener(IAsyncFetchListener listener) {
@@ -54,7 +61,11 @@ class DownloadFile extends AsyncTask<List<Object>, Integer, String> {
 
 		Integer filesCount = fileToDownload[0].size();
 		Integer i = 0;
+
 		File fileObj = null;
+		InputStream input = null;
+		OutputStream output = null;
+		HttpURLConnection connection = null;
 
 		try {
 
@@ -63,15 +74,7 @@ class DownloadFile extends AsyncTask<List<Object>, Integer, String> {
 				HashMap<String, String> hashMap = (HashMap<String, String>) fileToDownload[0]
 						.get(i);
 
-				int count;
-
 				String strUrl = hashMap.get("url");
-				URL url = new URL(strUrl);
-				URLConnection connection = url.openConnection();
-				connection.connect();
-
-				int lengthOfFile = connection.getContentLength();
-				long total = 0;
 				String name;
 				if (hashMap.containsKey("name")) {
 					name = hashMap.get("name");
@@ -79,23 +82,40 @@ class DownloadFile extends AsyncTask<List<Object>, Integer, String> {
 					name = strUrl.substring(strUrl.lastIndexOf('/'));
 				}
 				fileObj = new File(outputDirectory, name);
-				InputStream input = new BufferedInputStream(url.openStream());
-				OutputStream output = new FileOutputStream(fileObj, false);
-				byte data[] = new byte[1024];
+				deleteUnCompletedFile(fileObj);
+
+				URL url = new URL(strUrl);
+				connection = (HttpURLConnection) url.openConnection();
+				connection.setUseCaches(useCache);
+				connection.connect();
+
+				if (connection.getResponseCode() != HttpURLConnection.HTTP_OK) {
+					Log.e("Server returned HTTP ", connection.getResponseCode()
+							+ " - " + connection.getResponseMessage());
+				}
+
+				int fileLength = connection.getContentLength();
+
+				input = connection.getInputStream();
+				output = new FileOutputStream(fileObj);
+
+				byte data[] = new byte[4096];
+				long total = 0;
+				int count;
 				while ((count = input.read(data)) != -1) {
 					total += count;
-					int fileProgress = (int) ((total * 100) / lengthOfFile);
-					publishProgress((int) ((fileProgress + (i * 100)) / filesCount));
+					if (fileLength > 0) {
+						int fileProgress = (int) ((total * 100) / fileLength);
+						publishProgress((int) ((fileProgress + (i * 100)) / filesCount));
+					}
 					output.write(data, 0, count);
 				}
-				output.flush();
-				output.close();
-				input.close();
 
 				if (isCancelled()) {
 					deleteUnCompletedFile(fileObj);
 					break;
 				}
+
 			}
 
 		} catch (Exception e) {
@@ -105,6 +125,25 @@ class DownloadFile extends AsyncTask<List<Object>, Integer, String> {
 				notificationHelper.notcompleted();
 			}
 			this.fetchListener.onError(e.toString(), i);
+
+		} finally {
+
+			try {
+
+				if (output != null) {
+					output.flush();
+					output.close();
+				}
+
+				if (input != null)
+					input.close();
+
+			} catch (IOException ignored) {
+
+			}
+
+			if (connection != null)
+				connection.disconnect();
 
 		}
 
@@ -134,9 +173,29 @@ class DownloadFile extends AsyncTask<List<Object>, Integer, String> {
 		this.fetchListener.onCancel();
 	}
 
+	private static SSLSocketFactory createSslSocketFactory() throws Exception {
+		TrustManager[] byPassTrustManagers = new TrustManager[] { new X509TrustManager() {
+			public X509Certificate[] getAcceptedIssuers() {
+				return new X509Certificate[0];
+			}
+
+			public void checkClientTrusted(X509Certificate[] chain,
+					String authType) {
+			}
+
+			public void checkServerTrusted(X509Certificate[] chain,
+					String authType) {
+			}
+		} };
+		SSLContext sslContext = SSLContext.getInstance("TLS");
+		sslContext.init(null, byPassTrustManagers, new SecureRandom());
+		return sslContext.getSocketFactory();
+	}
+
 	private void deleteUnCompletedFile(File file) {
 		try {
-			file.delete();
+			if (file.exists())
+				file.delete();
 		} catch (Exception err) {
 			// error
 		}
